+++ spreadsheet-util/src/main/java/module-info.java
@@ -23,20 +23,24 @@ module nbbrd.spreadsheet.util {
     exports shaded.spreadsheet.nbbrd.io to
             nbbrd.spreadsheet.html,
             nbbrd.spreadsheet.poi,
-            nbbrd.spreadsheet.xmlss;
+            nbbrd.spreadsheet.xmlss,
+            nbbrd.spreadsheet.xl;
 
     exports shaded.spreadsheet.nbbrd.io.function to
             nbbrd.spreadsheet.html,
             nbbrd.spreadsheet.poi,
-            nbbrd.spreadsheet.xmlss;
+            nbbrd.spreadsheet.xmlss,
+            nbbrd.spreadsheet.xl;
 
     exports shaded.spreadsheet.nbbrd.io.xml to
             nbbrd.spreadsheet.html,
             nbbrd.spreadsheet.poi,
-            nbbrd.spreadsheet.xmlss;
+            nbbrd.spreadsheet.xmlss,
+            nbbrd.spreadsheet.xl;
 
     exports shaded.spreadsheet.nbbrd.io.zip to
             nbbrd.spreadsheet.html,
             nbbrd.spreadsheet.poi,
-            nbbrd.spreadsheet.xmlss;
+            nbbrd.spreadsheet.xmlss,
+            nbbrd.spreadsheet.xl;
 }